simplified homepage code by using components
