+++ client/src/components/Footer.js
@@ -1,0 +1,15 @@
+function Footer() {
+  return (
+    <footer class="p-4 bg-white shadow md:px-6 md:py-8 dark:bg-gray-900">
+      <span class="block text-sm text-gray-500 sm:text-center dark:text-gray-400">
+        © 2022{' '}
+        <a href="https://flowbite.com/" class="hover:underline">
+          identiFun™
+        </a>
+        . All Rights Reserved.
+      </span>
+    </footer>
+  );
+}
+
+export default Footer;
